# alzuhud.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "alzuhud services",
    "version": "1.1.0"
  },
  "host": "165.22.234.156:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/auth/refresh": {
      "get": {
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "refresh token",
            "schema": {
              "$ref": "#/definitions/AlzuhudResponseAuth"
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "description": "if you use this api must use new token before token expired.",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "tags": [
          "Alzuhud"
        ]
      }
    },
...
    "/api/guest/contact/add": {
      "post": {
        "operationId": "contact",
        "responses": {
          "200": {
            "description": "contact",
            "schema": {
              "$ref": "#/definitions/AlzuhudResponseVoid"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AlzuhudRequestAddContact"
            }
          }
        ],
        "tags": [
          "Alzuhud"
        ]
      }
    },
    "/api/admin/contact/all": {
      "get": {
        "operationId": "all contacts",
        "responses": {
          "200": {
            "description": "contacts",
            "schema": {
              "$ref": "#/definitions/AlzuhudResponseContacts"
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Alzuhud"
        ]
      }
    },
...
  },
  "definitions": {
...
    "AlzuhudRequestAddContact": {
      "type": "object",
      "properties": {
        "subject": {
          "type": "string",
          "description": "required and max length is 50"
        },
        "message": {
          "type": "string",
          "description": "required and max length is 150000"
        },
        "email": {
          "type": "string",
          "format": "email",
          "description": "optional and max length is 200 and must be a valid email address"
        },
        "name": {
          "type": "string",
          "description": "optional and max length is 50"
        }
      }
    },
...
    "AlzuhudResponseVoid": {
      "type": "object"
    },
    "AlzuhudResponseContacts": {
      "type": "object",
      "properties": {
        "current_page": {
          "type": "integer"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "user_id": {
                "type": "object",
                "items": {
                  "$ref": "#/definitions/UserObjectDefintion"
                }
              },
              "subject": {
                "type": "string",
                "format": "string"
              },
              "body": {
                "type": "string",
                "format": "string"
              },
              "status": {
                "type": "string",
                "format": "string"
              },
              "ticket_img": {
                "type": "string",
                "format": "base64"
              },
              "ticket_img_url": {
                "type": "string",
                "format": "url"
              },
              "time_ago": {
                "type": "string",
                "format": "string"
              },
              "human_created_at": {
                "type": "string",
                "format": "string"
              }
            }
          }
        },
        "first_page_url": {
          "type": "string",
          "format": "string"
        },
        "from": {
          "type": "integer",
          "format": "int64"
        },
        "last_page": {
          "type": "integer",
          "format": "int64"
        },
        "last_page_url": {
          "type": "string",
          "format": "url"
        },
        "next_page_url": {
          "type": "string",
          "format": "url"
        },
        "path": {
          "type": "string",
          "format": "uri"
        },
        "per_page": {
          "type": "integer",
          "format": "int32"
        },
        "prev_page_url": {
          "type": "integer",
          "format": "int32"
        },
        "to": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "UserObjectDefintion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "full_name": {
          "type": "string",
          "format": "string"
        },
        "email": {
          "type": "string",
          "format": "string"
        },
        "phone": {
          "type": "string",
          "format": "string"
        },
        "role": {
          "type": "string",
          "format": "string"
        },
        "status": {
          "type": "string",
          "format": "string"
        },
        "email_verified_at": {
          "type": "string",
          "format": "timestamp"
        },
        "remember_token": {
          "type": "string",
          "format": "timestamp"
        },
        "created_at": {
          "type": "string",
          "format": "string"
        },
        "updated_at": {
          "type": "string",
          "format": "timestamp"
        }
      }
    },
...
    "AlzuhudResponseAuth": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "token_type": {
          "type": "string"
        },
        "expires_in": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "items": {
            "$ref": "#/definitions/UserObjectDefintion"
          }
        }
      }
    },
...
  }
}